% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hi.R
\name{hi}
\alias{hi}
\title{Helps start conversations}
\usage{
hi(who, how = c("shout", "whisper"))
}
\arguments{
\item{who}{character(1) The name of the person you wish to talk to start a
conversation with}

\item{how}{char(1) how to greet the person.  Either "shout" (default) or
"whisper"}
}
\value{
character(1) A line of hello world with num of chars
}
\description{
This function generates a string with name, numchars ldladnflanflanfnal
ladkflfamlaklallafmkaaflmlamalfk
}
\examples{
hi("Paul")
hi("Paul", "whisper")

}

