% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rich.R
\name{input_rich}
\alias{input_rich}
\title{Rich data input}
\usage{
input_rich(pdata_file, exprs_file)
}
\arguments{
\item{pdata_file}{character (1) The path to the pheno data file.}

\item{exprs_file}{Character (1) The path to the expression file.}
}
\value{
A summarized experiment containing samples as colData(), and 
the expression values as assay()
}
\description{
Capture semantic meaning of data we represent.  For instance, separately 
 recognizing the phenotypic data describing sample and the expression
 values derives from microarray or RNAseq experiment.
}

