% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_unit_convert.R
\name{MPa_to_PSI}
\alias{MPa_to_PSI}
\title{Stress Unit Conversion from MPa to psi}
\usage{
MPa_to_PSI(MPa)
}
\arguments{
\item{MPa}{The stress in MPa}
}
\value{
The stress in psi
}
\description{
Convert stress in MPa to stress in psi
}
\examples{
stress1 <- MPa_to_PSI(0.7);
stress2 <- MPa_to_PSI(c(0.4, 0.6, 0.8));
}
