% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stress_unit_convert.R
\name{PSI_to_MPa}
\alias{PSI_to_MPa}
\title{Stress Unit Conversion from psi to MPa}
\usage{
PSI_to_MPa(PSI)
}
\arguments{
\item{PSI}{The stress in psi}
}
\value{
The stress in MPa
}
\description{
Convert stress in psi to stress in MPa
}
\examples{
stress1 <- PSI_to_MPa(100);
stress2 <- PSI_to_MPa(c(40,60,800));
}
